將 postgres 表示 id 的欄位由 serial 調整爲 bigserial

+++ src/db/models.rs
@@ -24,7 +24,7 @@ pub struct NewUser<'a> {
 
 #[derive(Queryable)]
 pub struct Invitation {
-    pub id: i32,
+    pub id: i64,
     pub code: String,
     pub email: String,
     pub create_time: std::time::SystemTime,
@@ -38,59 +38,59 @@ pub struct NewInvitation<'a> {
 
 #[derive(Queryable)]
 pub struct Board {
-    pub id: i32,
+    pub id: i64,
     pub board_name: String,
-    pub ruling_party_id: i32,
+    pub ruling_party_id: i64,
 }
 
 #[derive(Insertable)]
 #[table_name = "boards"]
 pub struct NewBoard<'a> {
     pub board_name: &'a str,
-    pub ruling_party_id: i32,
+    pub ruling_party_id: i64,
 }
 
 #[derive(Queryable)]
 pub struct NodeTemplate {
-    pub id: i32,
-    pub board_id: i32,
+    pub id: i64,
+    pub board_id: i64,
     pub def: String,
 }
 #[derive(Insertable)]
 #[table_name = "node_templates"]
 pub struct NewNodeTemplate {
-    pub board_id: i32,
+    pub board_id: i64,
     pub def: String,
 }
 
 #[derive(Queryable)]
 pub struct Edge {
-    pub id: i32,
-    pub from_node: i32,
-    pub to_node: i32,
-    pub transfuse: i32,
+    pub id: i64,
+    pub from_node: i64,
+    pub to_node: i64,
+    pub transfuse: i64,
 }
 #[derive(Insertable)]
 #[table_name = "edges"]
 pub struct NewEdge {
-    pub from_node: i32,
-    pub to_node: i32,
+    pub from_node: i64,
+    pub to_node: i64,
     pub transfuse: i32,
 }
 
 #[derive(Queryable)]
 pub struct Article {
-    pub id: i32,
-    pub board_id: i32,
-    pub template_id: i32,
+    pub id: i64,
+    pub board_id: i64,
+    pub template_id: i64,
     pub author_id: String,
     pub article_name: String,
 }
 #[derive(Insertable)]
 #[table_name = "articles"]
 pub struct NewArticle {
-    pub board_id: i32,
-    pub template_id: i32,
+    pub board_id: i64,
+    pub template_id: i64,
     pub author_id: String,
     pub article_name: String,
 }
